+++ internal/interpreter/core/core_test.go
