+++ internal/interpreter/core/dependency.go
