+++ internal/interpreter/core/preprocessor.go
