+++ internal/interpreter/core/preprocessor_dependency.go
