+++ internal/interpreter/core/testdata/imports/fileC.txt
