+++ internal/interpreter/core/testdata/test.txt
