+++ internal/interpreter/core/token.go
