+++ internal/interpreter/core/variable.go
