+++ internal/interpreter/functions/crypt.go
