+++ internal/interpreter/functions/internal.go
