+++ internal/interpreter/functions/math.go
