+++ internal/interpreter/functions/string.go
