+++ internal/interpreter/functions/time.go
