+++ internal/interpreter/testdata/ignore/gold/ignore.yaml
