+++ internal/interpreter/testdata/imports/in/myTestImportOk.txt
