+++ internal/interpreter/testdata/interpret/in/partials/varscopes.yaml
